// web/src/lib/api/generated/models/poskasir-internal-dto-profile-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * POS Kasir API
 * POS Kasir API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { POSKasirInternalRepositoryUserRole } from './poskasir-internal-repository-user-role';

export interface POSKasirInternalDtoProfileResponse {
    'avatar'?: string;
    'created_at'?: string;
    'deleted_at'?: string;
    'email'?: string;
    'id'?: string;
    'is_active'?: boolean;
    'role'?: POSKasirInternalRepositoryUserRole;
    'updated_at'?: string;
    'username'?: string;
}